工具类编写和注册接口初始化，完善参数校验，增加了参数校验的处理异常

+++ src/main/java/cool/oriental/chatcove/entity/ChannelInfo.java
@@ -50,7 +50,7 @@ public class ChannelInfo implements Serializable {
      * 创建者id
      */
     @TableField("master_id")
-    private Integer masterId;
+    private Long masterId;
 
     /**
      * 频道描述
